refactor: upgrade useCreateExperimentModal to use useModal pattern
